Refactor back-end code

+++ frontend/src/components/AdminDashboard/utils/handleRemoveContributor.js
@@ -1,4 +1,5 @@
 const BACKEND_URL = import.meta.env.VITE_BACKEND_URL;
+import { getCsrfToken } from "../../utils/getCsrfToken";
 
 async function handleRemoveContributor(
   user,
@@ -6,13 +7,24 @@ async function handleRemoveContributor(
   addNotification,
 ) {
   try {
+    const csrfToken = await getCsrfToken();
+
+    if (!csrfToken) {
+      addNotification({
+        type: "error",
+        message: "Failed to retrieve CSRF token.",
+      });
+      return;
+    }
+
     const response = await fetch(
-      `${BACKEND_URL}/admin/remove-contributor/${user.id}`,
+      `${BACKEND_URL}/users/admin/remove-contributor/${user.id}`,
       {
         method: "DELETE",
         mode: "cors",
         headers: {
           "Content-Type": "application/json",
+          "x-csrf-token": csrfToken,
         },
         credentials: "include",
       },
